feed_news/requirements.txt
gspread==6.1.4
google-auth==2.38.0
requests==2.32.3
python-dotenv==1.1.0
pytz==2024.2
